+++ src/rendering/tile_map_renderer.cpp
@@ -26,7 +26,7 @@ void TileMapRenderer::draw(
             int frameIndex = tile.getCurrentFrame();
             auto [uvStart, uvEnd] = tileSet.getUVRange(frameIndex, tileSize);
             glm::vec2 position = tileMap.tileToWorldPosition(glm::ivec2(tileX, tileY));
-            glm::vec2 size = glm::vec2(tileSize);
+            glm::vec2 size = glm::vec2(static_cast<float>(tileSize));
 
             spriteRenderer.drawWithUV(tileSetShader, tileSet, projection, position, size, uvStart, uvEnd);
         }
